Doc2Md/requirement.txt
docling==2.18.0
gradio==5.14.0
grpcio-status==1.70.0
marker-pdf==1.3.5
multiprocess==0.70.17
openai==1.61.1
pipdeptree==2.25.0
pytesseract==0.3.13
semchunk==2.2.2
tesseract==0.1.3
